pyproject.toml
[tool.poetry]
name = "yolo-detector-ui"
version = "0.1.0"
description = ""
authors = ["joypauls <joypaulsen3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.9.0.80"
streamlit = "^1.30.0"
pillow = "^10.2.0"
torch = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
